Cloud Test/backend/requirements.txt
fastapi
uvicorn
pydantic
pinecone
openai
python-dotenv
tika
